Move TLSConfig to a common package
