+++ config/configgrpc/configgrpc_test.go
@@ -15,12 +15,11 @@
 package configgrpc
 
 import (
-	"crypto/x509"
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
+
+	"github.com/open-telemetry/opentelemetry-collector/config/configtls"
 )
 
 func TestBasicGrpcSettings(t *testing.T) {
@@ -29,7 +28,7 @@ func TestBasicGrpcSettings(t *testing.T) {
 		Headers:     nil,
 		Endpoint:    "",
 		Compression: "",
-		TLSConfig: TLSConfig{
+		TLSConfig: configtls.TLSConfig{
 			CaCert:             "",
 			UseSecure:          false,
 			ServerNameOverride: "",
@@ -51,7 +50,7 @@ func TestInvalidPemFile(t *testing.T) {
 				Headers:     nil,
 				Endpoint:    "",
 				Compression: "",
-				TLSConfig: TLSConfig{
+				TLSConfig: configtls.TLSConfig{
 					CaCert:             "/doesnt/exist",
 					UseSecure:          false,
 					ServerNameOverride: "",
@@ -65,7 +64,7 @@ func TestInvalidPemFile(t *testing.T) {
 				Headers:     nil,
 				Endpoint:    "",
 				Compression: "",
-				TLSConfig: TLSConfig{
+				TLSConfig: configtls.TLSConfig{
 					CaCert:             "/doesnt/exist",
 					UseSecure:          true,
 					ServerNameOverride: "",
@@ -79,7 +78,7 @@ func TestInvalidPemFile(t *testing.T) {
 				Headers:     nil,
 				Endpoint:    "",
 				Compression: "",
-				TLSConfig: TLSConfig{
+				TLSConfig: configtls.TLSConfig{
 					ClientCert:         "/doesnt/exist",
 					UseSecure:          true,
 					ServerNameOverride: "",
@@ -101,7 +100,7 @@ func TestUseSecure(t *testing.T) {
 		Headers:     nil,
 		Endpoint:    "",
 		Compression: "",
-		TLSConfig: TLSConfig{
+		TLSConfig: configtls.TLSConfig{
 			CaCert:             "",
 			UseSecure:          true,
 			ServerNameOverride: "",
@@ -111,122 +110,4 @@ func TestUseSecure(t *testing.T) {
 
 	assert.NoError(t, err)
 	assert.Equal(t, len(dialOpts), 1)
-}
-
-func TestOptionsToConfig(t *testing.T) {
-	tests := []struct {
-		name        string
-		options     TLSConfig
-		fakeSysPool bool
-		expectError string
-	}{
-		{
-			name:    "should load system CA",
-			options: TLSConfig{CaCert: ""},
-		},
-		{
-			name:        "should fail with fake system CA",
-			fakeSysPool: true,
-			options:     TLSConfig{CaCert: ""},
-			expectError: "fake system pool",
-		},
-		{
-			name:    "should load custom CA",
-			options: TLSConfig{CaCert: "testdata/testCA.pem"},
-		},
-		{
-			name:        "should fail with invalid CA file path",
-			options:     TLSConfig{CaCert: "testdata/not/valid"},
-			expectError: "failed to load CA",
-		},
-		{
-			name:        "should fail with invalid CA file content",
-			options:     TLSConfig{CaCert: "testdata/testCA-bad.txt"},
-			expectError: "failed to parse CA",
-		},
-		{
-			name: "should load valid TLS Client settings",
-			options: TLSConfig{
-				CaCert:     "testdata/testCA.pem",
-				ClientCert: "testdata/test-cert.pem",
-				ClientKey:  "testdata/test-key.pem",
-			},
-		},
-		{
-			name: "should fail with missing TLS Client Key",
-			options: TLSConfig{
-				CaCert:     "testdata/testCA.pem",
-				ClientCert: "testdata/test-cert.pem",
-			},
-			expectError: "both client certificate and key must be supplied",
-		},
-		{
-			name: "should fail with invalid TLS Client Key",
-			options: TLSConfig{
-				CaCert:     "testdata/testCA.pem",
-				ClientCert: "testdata/test-cert.pem",
-				ClientKey:  "testdata/not/valid",
-			},
-			expectError: "failed to load server TLS cert and key",
-		},
-		{
-			name: "should fail with missing TLS Client Cert",
-			options: TLSConfig{
-				CaCert:    "testdata/testCA.pem",
-				ClientKey: "testdata/test-key.pem",
-			},
-			expectError: "both client certificate and key must be supplied",
-		},
-		{
-			name: "should fail with invalid TLS Client Cert",
-			options: TLSConfig{
-				CaCert:     "testdata/testCA.pem",
-				ClientCert: "testdata/not/valid",
-				ClientKey:  "testdata/test-key.pem",
-			},
-			expectError: "failed to load server TLS cert and key",
-		},
-		{
-			name: "should fail with invalid TLS Client CA",
-			options: TLSConfig{
-				CaCert: "testdata/not/valid",
-			},
-			expectError: "failed to load CA",
-		},
-		{
-			name: "should fail with invalid Client CA pool",
-			options: TLSConfig{
-				CaCert: "testdata/testCA-bad.txt",
-			},
-			expectError: "failed to parse CA",
-		},
-		{
-			name: "should pass with valid Client CA pool",
-			options: TLSConfig{
-				CaCert: "testdata/testCA.pem",
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			if test.fakeSysPool {
-				saveSystemCertPool := systemCertPool
-				systemCertPool = func() (*x509.CertPool, error) {
-					return nil, fmt.Errorf("fake system pool")
-				}
-				defer func() {
-					systemCertPool = saveSystemCertPool
-				}()
-			}
-			cfg, err := test.options.LoadTLSConfig()
-			if test.expectError != "" {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), test.expectError)
-			} else {
-				require.NoError(t, err)
-				assert.NotNil(t, cfg)
-			}
-		})
-	}
 }
