+++ exporter/opencensusexporter/factory_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector/compression"
 	"github.com/open-telemetry/opentelemetry-collector/config/configcheck"
 	"github.com/open-telemetry/opentelemetry-collector/config/configgrpc"
+	"github.com/open-telemetry/opentelemetry-collector/config/configtls"
 	"github.com/open-telemetry/opentelemetry-collector/exporter/exportertest"
 	"github.com/open-telemetry/opentelemetry-collector/receiver/opencensusreceiver"
 	"github.com/open-telemetry/opentelemetry-collector/testutils"
@@ -88,7 +89,7 @@ func TestCreateTraceExporter(t *testing.T) {
 			config: Config{
 				GRPCSettings: configgrpc.GRPCSettings{
 					Endpoint: rcvCfg.Endpoint,
-					TLSConfig: configgrpc.TLSConfig{
+					TLSConfig: configtls.TLSConfig{
 						UseSecure: true,
 					},
 				},
@@ -161,7 +162,7 @@ func TestCreateTraceExporter(t *testing.T) {
 			config: Config{
 				GRPCSettings: configgrpc.GRPCSettings{
 					Endpoint: rcvCfg.Endpoint,
-					TLSConfig: configgrpc.TLSConfig{
+					TLSConfig: configtls.TLSConfig{
 						CaCert: "testdata/test_cert.pem",
 					},
 				},
@@ -172,7 +173,7 @@ func TestCreateTraceExporter(t *testing.T) {
 			config: Config{
 				GRPCSettings: configgrpc.GRPCSettings{
 					Endpoint: rcvCfg.Endpoint,
-					TLSConfig: configgrpc.TLSConfig{
+					TLSConfig: configtls.TLSConfig{
 						CaCert: "nosuchfile",
 					},
 				},
